finished simplifying queries API

+++ classes/glue/query/delete.php
@@ -1,17 +1,8 @@
 <?php defined('SYSPATH') OR die('No direct access allowed.');
 
 class Glue_Query_Delete extends Glue_Query {
-	public function __construct($entity_name, &$set, $conditions = null) {
-		parent::__construct($entity_name, $set, $conditions);
-
-		// Load only pk :
-		$this->active_command->fields($set->entity()->pk());
-
-		return $this;
-	}
-
-	public function with($src_set, $relationship, &$trg_set = null) {
-		parent::with($src_set, $relationship, $trg_set);
+	public function __construct($entity_name, &$trg_set = null, $conditions = null, $order_by = null, $limit = null, $offset = null) {
+		parent::__construct($entity_name, $trg_set, $conditions, $order_by, $limit, $offset);
 
 		// Load only pk :
 		$this->active_command->fields($trg_set->entity()->pk());
@@ -19,11 +10,20 @@ class Glue_Query_Delete extends Glue_Query {
 		return $this;
 	}
 
-	public function execute() {
-		parent::execute();
+	public function with($src_set, $relationship_name, &$trg_set = null, $conditions = null, $order_by = null, $limit = null, $offset = null) {
+		parent::with($src_set, $relationship_name, $trg_set, $conditions, $order_by, $limit, $offset);
 
-		// Delete set contents :
-		foreach($this->sets as $hash => $data)
-			$data['set']->delete();
+		// Load only pk :
+		$this->active_command->fields($trg_set->entity()->pk());
+
+		return $this;
+	}
+
+	public function exec() {
+		if (parent::exec()) {
+			// Delete set contents :
+			foreach($this->sets as $hash => $data)
+				$data['set']->delete();
+		}
 	}
 }
